Refactor each session type into a separate module
This is a purely internal change, to make it easier to compare and
understand the implementation of each primitive.
